aws/requirements.txt
setuptools
Flask
flask-cors
flask_limiter
waitress
folium
pyproj
GitPython
git+https://github.com/urasakikeisuke/pyflann.git
